This commit was manufactured by cvs2svn to create tag 'wxPy_2_6_3_3'.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/tags/wxPy_2_6_3_3@40115 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/variant.cpp
@@ -473,7 +473,7 @@ bool wxVariantDataReal::Write(wxSTD ostream& str) const
 
 bool wxVariantDataReal::Write(wxString& str) const
 {
-    str.Printf(wxT("%.4f"), m_value);
+    str.Printf(wxT("%.14g"), m_value);
     return true;
 }
 
@@ -709,7 +709,7 @@ bool wxVariantDataChar::Read(wxInputStream& str)
 
 bool wxVariantDataChar::Read(wxString& str)
 {
-    m_value = str.ToAscii()[0u];
+    m_value = str.ToAscii()[size_t(0)];
     return true;
 }
